Delete Agents folder and create a new file to store them
